made the centroid a float
git-svn-id: svn://orca.physics.unc.edu/Orca@927 fd86f90f-2c0a-422a-b972-84106d8db31b

+++ Source/Widgets/Plotting/ORGate1D.m
@@ -365,7 +365,7 @@ const float kGateAlpha2 = .1;
 
 - (float) gatePeakValue
 {
-	int val = [self peakx];
+	float val = [self peakx];
 	id theCalibration = [[mPlot xScale] calibration];
 	if(theCalibration && [theCalibration useCalibration])return [theCalibration convertedValueForChannel:val];
 	else return val;
@@ -373,7 +373,7 @@ const float kGateAlpha2 = .1;
 
 - (float) gateCentroid
 {
-	int val = [self centroid];
+	float val = [self centroid];
 	id theCalibration = [[mPlot xScale] calibration];
 	if(theCalibration && [theCalibration useCalibration])return [theCalibration convertedValueForChannel:val];
 	else return val;
@@ -381,7 +381,7 @@ const float kGateAlpha2 = .1;
 
 - (float) gateSigma
 {
-	int val = [self sigma];
+	float val = [self sigma];
 	id theCalibration = [[mPlot xScale] calibration];
 	if(theCalibration && [theCalibration useCalibration])return [theCalibration convertedValueForChannel:val];
 	else return val;
